[#46] Replace PgSharedService with utils functions
