--- src/services/postgres/postgres.shared.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PgSharedService } from './postgres.service';
-
-@Module({
-  providers: [PgSharedService],
-  exports: [PgSharedService]
-})
-export class PostgresSharedModule {}
