+++ src/services/uploader/uploader.shared.module.ts
@@ -2,16 +2,14 @@ import { Module } from '@nestjs/common';
 import { UploaderService } from './uploader.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UploaderEntity } from '@app-entities';
-import { PgSharedService } from '../postgres/postgres.service';
 import { UserSharedModule } from '../user/user.shared.module';
 
 @Module({
   imports: [
-    PgSharedService,
     UserSharedModule,
     TypeOrmModule.forFeature([UploaderEntity])
   ],
-  providers: [PgSharedService, UploaderService],
+  providers: [UploaderService],
   exports: [UploaderService],
 })
 export class UploaderSharedModule {}
