+++ src/services/user/user.shared.module.ts
@@ -2,11 +2,10 @@ import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserEntity } from '@app-entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { PgSharedService } from '../postgres/postgres.service';
 
 @Module({
-  providers: [UserService, PgSharedService],
-  imports: [PgSharedService, TypeOrmModule.forFeature([UserEntity])],
+  providers: [UserService],
+  imports: [TypeOrmModule.forFeature([UserEntity])],
   exports: [UserService],
 })
 export class UserSharedModule {}
